Move ListView to pkg/widgets and remove pager opened/closed
